[WIP] Move client and publicClient to useApolloClient hook
